Working design for solveInPlace for SparseElimination
