+++ linbox/solutions/solve-wip.h
@@ -29,6 +29,9 @@
 #include <linbox/solutions/methods-wip.h>
 
 namespace LinBox {
+    //
+    // solve
+    //
 
     /**
      * \brief Solve Ax = b, for x.
@@ -212,10 +215,90 @@ namespace LinBox {
     {
         solve(xNum, xDen, A, b, MethodWIP::Auto());
     }
+
+    //
+    // solveInPlace
+    //
+
+    /**
+     * \brief Solve Ax = b, for x.
+     *
+     * Returns a vector x such that Ax = b.
+     * A can be modified.
+     *
+     * See documentation for `solve`.
+     */
+    template <class ResultVector, class Matrix, class Vector, class CategoryTag, class SolveMethod>
+    inline ResultVector& solveInPlace(ResultVector& x, Matrix& A, const Vector& b, const CategoryTag& tag, const SolveMethod& m)
+    {
+        // @note This is called if the specialization has not been implemented,
+        // which means there might not be any "in place" version of the solve.
+        // So, we just forward to classic solve.
+
+        return solve(x, A, b, tag, m);
+    }
+
+    /**
+     * \brief Solve in place dispatcher for automated category tag.
+     */
+    template <class ResultVector, class Matrix, class Vector, class SolveMethod>
+    inline ResultVector& solveInPlace(ResultVector& x, Matrix& A, const Vector& b, const SolveMethod& m)
+    {
+        return solveInPlace(x, A, b, typename FieldTraits<typename Matrix::Field>::categoryTag(), m);
+    }
+
+    /**
+     * \brief Solve in place dispatcher for automated solve method.
+     */
+    template <class ResultVector, class Matrix, class Vector>
+    inline ResultVector& solveInPlace(ResultVector& x, Matrix& A, const Vector& b)
+    {
+        return solveInPlace(x, A, b, MethodWIP::Auto());
+    }
+
+    /**
+     * \brief Rational solve in place Ax = b, for x expressed as xNum/xDen.
+     * The matrix A might be modified.
+     *
+     * Second interface for solving in place, only valid for RingCategories::IntegerTag.
+     */
+    template <class Matrix, class Vector, class SolveMethod, class CategoryTag>
+    inline void solveInPlace(Vector& xNum, typename Vector::Field::Element& xDen, Matrix& A, const Vector& b,
+                             const CategoryTag& tag, const SolveMethod& m)
+    {
+        // @note This is called if the specialization has not been implemented,
+        // which means there might not be any "in place" version of the solve.
+        // So, we just forward to classic solve.
+
+        solve(xNum, xDen, A, b, tag, m);
+    }
+
+    /**
+     * \brief Rational solve in place dispatcher for automated category tag.
+     */
+    template <class Matrix, class Vector, class SolveMethod>
+    inline void solveInPlace(Vector& xNum, typename Vector::Field::Element& xDen, Matrix& A, const Vector& b,
+                             const SolveMethod& m)
+    {
+        solveInPlace(xNum, xDen, A, b, typename FieldTraits<typename Matrix::Field>::categoryTag(), m);
+    }
+
+    /**
+     * \brief Rational solve in place dispatcher for automated solve method.
+     */
+    template <class Matrix, class Vector>
+    inline void solveInPlace(Vector& xNum, typename Vector::Field::Element& xDen, Matrix& A, const Vector& b)
+    {
+        solveInPlace(xNum, xDen, A, b, MethodWIP::Auto());
+    }
 }
 
+//
+// solve
+//
+
 #include "./solve/solve-auto.h"
-// #include "./solve/solve-blackbox.h"
+// #include "./solve/solve-blackbox.h" @fixme Implement
 
 // Elimination
 #include "./solve/solve-dense-elimination.h"
@@ -230,7 +313,5 @@ namespace LinBox {
 // Blackbox
 #include "./solve/solve-wiedemann.h"
 #include "./solve/solve-lanczos.h"
-
-// #include "./solve/solvein.h" @fixme
 
 #endif
