+++ tests/test-solve-wip.C
@@ -33,6 +33,7 @@ void run_integer(Communicator& communicator, size_t dimension) {
 
     try {
         solve(x, A, b, method);
+        solveInPlace(x, A, b, method);
     } catch (...) {
         std::cout << "===> OUCH: Throwing error" << std::endl;
         return;
@@ -71,6 +72,7 @@ void run_modular() {
 
     try {
         solve(x, A, b, method);
+        solveInPlace(x, A, b, method);
     } catch (...) {
         std::cout << "===> OUCH: Throwing error" << std::endl;
         return;
@@ -92,7 +94,7 @@ int main(void)
 
     commentator().setReportStream(std::cout);
 
-    // run_integer<Givaro::ZRing<Integer>, MethodWIP::Auto>(communicator, 2);
+    run_integer<Givaro::ZRing<Integer>, MethodWIP::Auto>(communicator, 2);
     // run_integer<Givaro::ZRing<Integer>, MethodWIP::CraAuto>(communicator, 2);
     // run_integer<Givaro::ZRing<Integer>, MethodWIP::CraAuto>(communicator, 3);
     // run_integer<Givaro::ZRing<Integer>, MethodWIP::Dixon>(communicator, 2);
@@ -102,15 +104,15 @@ int main(void)
     // run_integer<Givaro::ZRing<Integer>, MethodWIP::NumericSymbolicNorm>(communicator, 2); // @fixme Fails
     // run_integer<Givaro::ZRing<Integer>, MethodWIP::NumericSymbolicNorm>(communicator, 3); // @fixme Fails
 
-    // run_modular<DenseMatrix<Givaro::Modular<double>>, MethodWIP::Auto>();
-    // run_modular<SparseMatrix<Givaro::Modular<double>>, MethodWIP::Auto>();
+    run_modular<DenseMatrix<Givaro::Modular<double>>, MethodWIP::Auto>();
+    run_modular<SparseMatrix<Givaro::Modular<double>>, MethodWIP::Auto>();
     // run_modular<DenseMatrix<Givaro::Modular<double>>, MethodWIP::DenseElimination>();
     // run_modular<SparseMatrix<Givaro::Modular<double>>, MethodWIP::DenseElimination>();
     // run_modular<DenseMatrix<Givaro::Modular<double>>, MethodWIP::SparseElimination>();
     // run_modular<SparseMatrix<Givaro::Modular<double>>, MethodWIP::SparseElimination>();
     // // run_modular<DenseMatrix<Givaro::Modular<double>>, MethodWIP::Wiedemann>(); // @fixme Can't compile
-    // run_modular<SparseMatrix<Givaro::Modular<double>>, MethodWIP::Wiedemann>();
-    run_modular<DenseMatrix<Givaro::Modular<double>>, MethodWIP::Lanczos>(); // @fixme Segmentation fault
+    run_modular<SparseMatrix<Givaro::Modular<double>>, MethodWIP::Wiedemann>();
+    // run_modular<DenseMatrix<Givaro::Modular<double>>, MethodWIP::Lanczos>(); // @fixme Segmentation fault
     // // run_modular<SparseMatrix<Givaro::Modular<double>>, MethodWIP::Lanczos>(); // @fixme Segmentation fault
     // // run_modular<DenseMatrix<Givaro::Modular<double>>, MethodWIP::BlockLanczos>(); // @fixme Can't compile
     // // run_modular<SparseMatrix<Givaro::Modular<double>>, MethodWIP::BlockLanczos>(); // @fixme Segmentation fault
